Remove View-Model contracts.
Remove Contracts class. Separe View interfaces and model classes from
contracts.
