+++ app/src/main/java/com/voltek/newsfeed/presentation/ui/list/ListView.kt
@@ -1,0 +1,11 @@
+package com.voltek.newsfeed.presentation.ui.list
+
+import com.arellomobile.mvp.viewstate.strategy.AddToEndSingleStrategy
+import com.arellomobile.mvp.viewstate.strategy.StateStrategyType
+import com.voltek.newsfeed.presentation.base.BaseView
+
+interface ListView : BaseView {
+
+    @StateStrategyType(AddToEndSingleStrategy::class)
+    fun render(model: ListModel)
+}
